[Manage roles] Refact waiting tx
